+++ server/utilities/middleware/logger.spec.js
@@ -1,4 +1,4 @@
-import { parseSuccessfulHealthChecks } from './logger';
+import { parseIgnoredRoutes } from './logger';
 import { expect } from 'chai';
 
 describe('Logger Middleware', () => {
@@ -13,7 +13,22 @@ describe('Logger Middleware', () => {
     };
 
     const expected = true;
-    const actual = parseSuccessfulHealthChecks(request, response);
+    const actual = parseIgnoredRoutes(request, response);
+    
+    expect(actual).equals(expected);
+  });
+
+  it('should NOT log the message if the root returns a 200 status', () => {
+
+    const request = {
+      url: '/'
+    };
+    const response = {
+      statusCode: 200
+    };
+
+    const expected = true;
+    const actual = parseIgnoredRoutes(request, response);
 
     expect(actual).equals(expected);
   });
@@ -28,13 +43,28 @@ describe('Logger Middleware', () => {
     };
 
     const expected = false;
-    const actual = parseSuccessfulHealthChecks(request, response);
+    const actual = parseIgnoredRoutes(request, response);
+    
+    expect(actual).equals(expected);
+  });
+
+  it('should log the message if the root returns a 400 status', () => {
+
+    const request = {
+      url: '/'
+    };
+    const response = {
+      statusCode: 400
+    };
+
+    const expected = false;
+    const actual = parseIgnoredRoutes(request, response);
 
     expect(actual).equals(expected);
   });
 
   it('should log the message if the url is not /healthz', () => {
-
+    
     const request = {
       url: '/foo'
     };
@@ -43,8 +73,9 @@ describe('Logger Middleware', () => {
     };
 
     const expected = false;
-    const actual = parseSuccessfulHealthChecks(request, response);
+    const actual = parseIgnoredRoutes(request, response);
 
     expect(actual).equals(expected);
   });
+
 });